Crate factor_graph

Crate allowing creation and manipulation of probabilistic factor graphs.

Reexports

pub use variable::Variable;

pub use variable::DiscreteVariable;

pub use factor::Factor;

pub use tree::SpanningTree;

pub use tree::TreeNode;

Modules

factor
Module holding factor-specific functionality

tree
Module with spanning tree code

variable
Module with variable-specific functionality

Structs

FactorGraph
Struct representing the full factor graph.

Traits

FactorGraphItem
Trait representing a generic item stored in the factor graph.