Struct factor_graph::FactorGraph

pub struct FactorGraph { /* fields omitted */ }

Struct representing the full factor graph.

Methods

impl FactorGraph

fn new() -> FactorGraph

Create an empty FactorGraph

fn add_discrete_var<T: Debug + Clone + 'static>(
&mut self,
name: &str,
val_names: Vec<T>
)

Add a new variable with the specified name to the factor graph.

fn add_factor<T: Debug + 'static>(
&mut self,
variables: Vec<String>,
func: fn(_: &[u32]) -> i32
)

Add a new factor with the specified variables to the factor graph.

fn render_to<W: Write>(&self, output: &mut W)

Render this graph to a Graphviz file

fn make_spanning_tree(&self, var: &str) -> SpanningTree

Make a spanning tree from the current factor graph with the specified root variable.

fn render_spanning_tree_to<W: Write>(&self, root_var: &str, output: &mut W)

Render a spanning tree for the factor graph to the input file, starting from the input variable.